+++ src/app/(mypage)/_components/MailboxSettingSection.tsx
